task 2 update2

+++ 2-do_deploy_web_static.py
@@ -7,9 +7,7 @@ Description: Fabric script (based on the file 1-pack_web_static.py that
 Author: a_idk
 """
 import os.path
-from fabric.api import env
-from fabric.api import put
-from fabric.api import run
+from fabric.api import env, put, run
 
 env.hosts = ['54.175.115.175', '54.237.55.177']  # my web servers
 
@@ -25,8 +23,34 @@ def do_deploy(archive_path):
     try:
         d_file = archive_path.split("/")[-1]
         f_name = d_file.split(".")[0]
-        f_path = "/data/web_static/releases/"
-        put(archive_path, '/tmp/')
+        # f_path = "/data/web_static/releases/"
+        if put(archive_path, f'/tmp/{d_file}').failed is True:
+            return False
+        if run(
+                f'rm -rf /data/web_static/releases/{f_name}/'
+                ).failed is True:
+            return False
+        if run(f'mkdir -p /data/web_static/releases/{f_name}/').failed is True:
+            return False
+        if run(f'tar -xzf /tmp/{d_file} -C /data/web_static/releases/{d_name}/'
+               ).failed is True:
+            return False
+        if run(f'rm /tmp/{d_file}').failed is True:
+            return False
+        if run(f'mv /data/web_static/releases/{f_name}/web_static/* '
+                f'/data/web_static/releases/{f_name}/').failed is True:
+            return False
+        if run(
+                f'rm -rf /data/web_static/releases/{f_name}/web_static'
+                ).failed is True:
+            return False
+        if run("rm -rf /data/web_static/current").failed is True:
+            return False
+        if run(f'ln -s /data/web_static/releases/{name}/ '
+                f'/data/web_static/current').failed is True:
+            return False
+        return True
+    """
         run(f'mkdir -p {f_path}{f_name}')
         run(f'tar -zxvf /tmp/{d_file} -C {f_path}{f_name}')
         run(f'rm /tmp/{d_file}')
@@ -37,3 +61,4 @@ def do_deploy(archive_path):
         return True
     except Exception as e:
         return False
+    """
